Update CMS files with new layout changes
