+++ site/content/first-time.md
@@ -40,7 +40,7 @@ cards:
 
       [Attend a Meetup »](https://events.chef.io)
   two:
-    title: Make your first contribution
+  title: Make your first contribution
     body: >-
       We always need help. Here are great places to start.
 
@@ -57,23 +57,7 @@ cards:
 
       [Help needed
       »](/help-wanted)
-swag_program:
-  body: >-
-    We’re glad you’re interested in contributing to a Chef project! We’re here
-    to do everything we can to make your first contribution experience a great
-    one.
-
-
-    We hope this guide will help you to choose where to contribute, and show you
-    the ropes of crafting your first contribution. If you have questions about
-    the development process, or are feeling stuck with your first contribution,
-    feel free to jump into specific project’s public development Slack channel,
-    or join a project-specific mailing list. The specific channels and mailing
-    list URLs can be found in the CONTRIBUTING.md file found in each project
-    repository.
-  header: Earn Chef swag for contributing
 ---
-
 # First time contributing?
 
 We’re glad you’re interested in contributing to a Chef project! We’re here to do everything we can to make your first contribution experience a great one.
