+++ site/content/products/chef-habitat.md
@@ -1,6 +1,6 @@
 ---
 title: Application Delivery as Code
-type: habitat
+type: chef-habitat
 hero: 
   logo: /img/logos/chef-habitat-dark.svg
 cards:
@@ -95,12 +95,4 @@ contact:
     link: https://www.chef.io/products/chef-habitat/
     copy: Learn more
 ---
-<p class="lead">Chef Habitat is an open source automation solution for defining, packaging, and delivering applications to almost any environment regardless of operating system or platform.</p>
-
-<div class="center-mobile">
-<a class="btn btn-primary" href="#learn">Get started</a>
-<a class="btn btn-secondary" href="https://bldr.habitat.sh/#/pkgs/core" target="_blank">Habitat Builder</a>
-</div>
-<iframe id="github-star" src="https://ghbtns.com/github-btn.html?user=habitat-sh&repo=habitat&type=star&count=true&size=large" frameborder="0" scrolling="0" width="160px" height="30px"></iframe>
-
-<a class="cta" href="https://discourse.chef.io/c/habitat/12" target="_blank">Release notes</a><br/>
+Test
