+++ src/admin/index.html
@@ -1,20 +1,22 @@
-<!doctype html>
-<html>
-<head>
-  <meta charset="utf-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Content Manager</title>
-   <script src="https://identity.netlify.com/v1/netlify-identity-widget.js"></script>
-   <link href="/admin/config.yml" type="text/yaml" rel="cms-config-url">
 
-</head>
-<body>
-  <!-- Include the script that builds the page and powers Netlify CMS -->
-  <script>
-  netlifyIdentity.init({
-    APIUrl: "https://community.chef.io/.netlify/identity"
-  });
-  </script>
-  
-</body>
+
+
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Content Manager</title>
+    <!-- Include the script that enables Netlify Identity on this page. -->
+    <script src="https://identity.netlify.com/v1/netlify-identity-widget.js"></script>
+    <link href="/admin/config.yml" type="text/yaml" rel="cms-config-url">
+  </head>
+  <body>
+    <!-- Include the script that builds the page and powers Netlify CMS -->
+    <script>
+      netlifyIdentity.init({
+        APIUrl: "https://community.chef.io/.netlify/identity"
+      });
+    </script>
+  </body>
 </html>
