+++ src/admin/netlify-cms/cms.js
@@ -2,16 +2,16 @@ import React from "react";
 import CMS from "netlify-cms-app";
 import Homepage from "./templates/homepage";
 import FirstTime from "./templates/first-time";
-import Inspec from "./templates/inspec";
-import Habitat from "./templates/habitat";
+import ChefInspec from "./templates/chef-inspec";
+import ChefHabitat from "./templates/chef-habitat";
 
 
 CMS.registerPreviewStyle('https://unpkg.com/netlify-cms-yoast-seo@~1.0/dist/main.css');
 
 CMS.registerPreviewTemplate("home", Homepage);
 CMS.registerPreviewTemplate("first-time", FirstTime);
-CMS.registerPreviewTemplate("inspec", Inspec);
-CMS.registerPreviewTemplate("habitat", Habitat);
+CMS.registerPreviewTemplate("chef-inspec", ChefInspec);
+CMS.registerPreviewTemplate("chef-habitat", ChefHabitat);
 
 
  // Register any CSS file on the home page as a preview style
