+++ src/admin/netlify-cms/inject-frontend-js.js
@@ -6,11 +6,11 @@ export function inject(){
 			const f = document.createElement("html");
 			f.innerHTML = html;
 			
-			if (document.querySelector(".css-1jl34xb-PreviewPaneFrame").contentDocument){
-				const iframedocument = document.querySelector(".css-1jl34xb-PreviewPaneFrame").contentDocument;
-			}else{
-				const iframedocument = document.querySelector(".css-1jl34xb-PreviewPaneFrame").contentWindow.document;
-			}
+			// if (document.querySelector(".css-1jl34xb-PreviewPaneFrame").contentDocument){
+			// 	const iframedocument = document.querySelector(".css-1jl34xb-PreviewPaneFrame").contentDocument;
+			// } else{
+			// 	const iframedocument = document.querySelector(".css-1jl34xb-PreviewPaneFrame").contentWindow.document;
+			// }
 			//const iframedocument = document.querySelector(".css-1jl34xb-PreviewPaneFrame").contentWindow.document;
 			//console.log(document.querySelector(".css-1jl34xb-PreviewPaneFrame"));
 			Array.from(f.getElementsByTagName("script")).forEach(tag => {
