+++ src/sass/custom.scss
@@ -974,8 +974,8 @@ pre {
 
 @import "marketo";
 @import "home";
-@import "inspec";
-@import "habitat";
+@import "chef-inspec";
+@import "chef-habitat";
 @import "first-time";
 @import "type";
 @import "markdown";
